requirements.txt
pymongo-4.13.2
colorama-0.4.6